+++ NEWS.md
@@ -1,5 +1,7 @@
-# ggbrain 0.8-1 (DEVELOPMENT)
+# ggbrain 0.8.1 (DEVELOPMENT)
 
+* feature: `ggbrain` objects can be added together, allowing common parts across plots to be reused.
+* feature: `geom_outline` now supports dilation and erosion using the `dil_ero` argument.
 * feature: `geom_region_label*` and `geom_region_text*` now support a `min_px` argument that controls
    the minimum number of pixels on a slice that will result in a label. Helps reduce labeling of small regions.
 * feature: support aes(outline=<x>, group=<y>) for geom_outline, allowing for subdivision of outlines